Connected client to lamda function #4
