+++ src/photoQueue/interfaces/ImageInterface.ts
