+++ src/photoQueue/interfaces/PhotoQueueProps.ts
