+++ src/testData.ts
@@ -1,9 +1,9 @@
-import { Image } from './interfaces/ImageInterface.ts';
 import cat1 from './assets/cat1.png';
 import cat2 from './assets/cat2.png';
 import cat3 from './assets/cat3.jpg';
 import cat4 from './assets/cat4.jpg';
 import cat5 from './assets/cat5.png';
+import { Image } from './photoQueue/interfaces/ImageInterface';
 
 export const testImages: Partial<Image>[] = [
   {
